requirements.txt
streamlit==1.41.1
python-dotenv==1.1.1
openai==1.109.1
langchain==0.3.0
langchain-openai==0.2.2
numpy==1.26.4
rich==13.9.4